# Conflicts:
#	MarfulApi/.vs/MarfulApi/DesignTimeBuild/.dtbcache.v2
#	MarfulApi/.vs/MarfulApi/v17/.futdcache.v1
#	MarfulApi/.vs/MarfulApi/v17/.suo
#	MarfulApi/MarfulApi/bin/Debug/net6.0/MarfulApi.dll
#	MarfulApi/MarfulApi/bin/Debug/net6.0/MarfulApi.pdb
#	MarfulApi/MarfulApi/bin/Debug/net6.0/ref/MarfulApi.dll
#	MarfulApi/MarfulApi/obj/Debug/net6.0/MarfulApi.csproj.AssemblyReference.cache
#	MarfulApi/MarfulApi/obj/Debug/net6.0/MarfulApi.csproj.CoreCompileInputs.cache
#	MarfulApi/MarfulApi/obj/Debug/net6.0/MarfulApi.csproj.FileListAbsolute.txt
#	MarfulApi/MarfulApi/obj/Debug/net6.0/MarfulApi.dll
#	MarfulApi/MarfulApi/obj/Debug/net6.0/MarfulApi.pdb
#	MarfulApi/MarfulApi/obj/Debug/net6.0/ref/MarfulApi.dll
#	MarfulApi/MarfulApi/obj/staticwebassets.pack.sentinel

+++ MarfulApi/MarfulApi/Data/BasketRepo.cs
@@ -15,17 +15,26 @@ namespace MarfulApi.Data
         {
             _db = db;
         }
-        public IQueryable<Basket> GetStudents => _db.Baskets;
+        public IQueryable<Basket> GetBaskets => _db.Baskets;
         public void Delete(int id)
         {
-            Basket basket = _db.Baskets.Find(id);
-            _db.Baskets.Remove(basket);
-            _db.SaveChanges();
+            var basket = _db.Baskets.First(p => p.Id == id);
+            if (basket != null)
+            {
+                _db.Baskets.Remove(basket);
+                _db.SaveChanges();
+            }
+            else
+                throw new FileNotFoundException();
+
         }
         public Basket GetBasket(int id)
         {
-            Basket basket = _db.Baskets.Find(id);
-            return basket;
+            var basket = _db.Baskets.First(p => p.Id == id);
+            if (basket != null)
+               return basket;
+            else
+                 throw new NotImplementedException();
         }
         public void Save(Basket basket)
         {
@@ -34,10 +43,16 @@ namespace MarfulApi.Data
                 _db.Baskets.Add(basket);
                 _db.SaveChanges();
             }
-            else
+           
+        }
+        public void Update(Basket basket)
+        {
+            var Basket = _db.Baskets.First(p => p.Id == basket.Id);
+            if(Basket != null)
             {
-                Basket Basket = _db.Baskets.Find(basket.Id);
+                Basket.ProductId = basket.ProductId;
                 Basket.TotalPrice = basket.TotalPrice;
+                Basket.UserPostId = basket.UserPostId;
                 _db.SaveChanges();
             }
         }
